Unify code options
Previously, multiple bools and strings were used to keep track of what
code should be added to the generated playground. Now, it’s all unified
into a single enum which is handled in a single place. This also makes
default framework imports unified for both URL and clipboard based code.