Add support for directory names with spaces
Adds the ability to use directory names that has spaces in it. Currently the playground is created successfully but opening it fails as it evaluates the path as two different arguments. This way we can save our playgrounds in iCloud Drive.

```
playground -t “/Users/[username]/Library/Mobile\ Documents/com~apple~CloudDocs/Playgrounds”
```